% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline_regression_basis_functions.R
\name{bspline_regression_basis_functions}
\alias{bspline_regression_basis_functions}
\title{Generate the basis functions needed for fitting the hazard B-spline model}
\usage{
bspline_regression_basis_functions(yd, entry, ORDER, knots, t)
}
\arguments{
\item{yd}{- matrix of time-to-event data; pneumonic y is event time, d is for delta, the status indicator}

\item{entry}{- late entry / left truncation times}

\item{ORDER}{- 1 step, 2 linear, 3 quadratic, 4 cubic}

\item{knots}{- sequence of knot locations}

\item{t}{- vector of evaluation times}
}
\value{
list(Wik, Zik, Eik, XH, Xh ) - see details for definitions
}
\description{
This function calculates all matrices needed for the estimation of a hazard function with censored time-to-event data using B-splines
}
\details{
When fitting a B-spline function \eqn{h(\alpha)(t) := B(t) \alpha} to time-to-event-data,
where \eqn{B(t)} is a basis of B-splines and \eqn{\alpha} are the coefficients to be estimated,
we optimise a function of the likelihood \deqn{L(h(\theta)) =  }.
}
