% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etsim_inputs_core.R
\name{etsim_inputs_core}
\alias{etsim_inputs_core}
\title{Core function used to simulate hazard and censoring distributions}
\usage{
etsim_inputs_core(
  Hazard_type = "spline",
  Param,
  Tmin = 0,
  Tmax = 10,
  SampleSize = 101
)
}
\arguments{
\item{Hazard_type}{- hazard function for outcome variable.
'exp' for exponential, 'weib' for Weibull, 'pe' for piecewise-exponential, or 'spline' for B-Spline}

\item{Param}{- matrix of parameter values for outcome variable}

\item{Tmin}{- scalar, minimum follow-up time, default 0}

\item{Tmax}{- scalar, maximum follow-up time, default 10}

\item{SampleSize}{- scalar, sample size, default 101}
}
\value{
Returns a list of outputs (t, h, Sh )
    t = array of times
    h = simulated hazard distribution
    Sh = cumulative hazard survival function
}
\description{
Generates hazard and censoring distributions and respective survival functions [IS IT OK TO IGNORE LATE ENTRY AT THIS STAGE?]
}
\details{
For 'exp', Parm contains a scalar;
    For 'pe', Parm(:, 1) is left limit, Parm(:, 2) is right limit, and Parm(:, 3) is hazard over the interval
    For 'bspline', Parm(:, 1) lists knots, Parm(:, 2) lists b-spline coefficients.
    The number of rows of Parm has to be equal to the number of degrees of freedom: number of interior knots plus order (=degree plus one) of the b-spline
}
\examples{
    #PENDING
    #Generate input for a cubic b-spline hazard simulation with piecewise exponential survival censoring
    knots = c(0, 1, 3, 6, 10, NaN, NaN )
    betac = 1 * c(0.05, 0.05, 0.05, 0.05, 0.40, 0.1, 0.05)
    HParm = data.frame(knots, betac) # 'A Simple B-Spline'
    INPUTS_hazard = etsim_inputs_core( HParam=HParm )

}
