% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline_basis.R
\name{generate_bspline_basis}
\alias{generate_bspline_basis}
\title{Generate B-Spline basis}
\usage{
generate_bspline_basis(time, Interior.knots, Boundary.knots, ORDER = 4)
}
\arguments{
\item{time}{- x coordinates for the B-spline functions}

\item{Interior.knots}{- interior knots of the B-spline basis}

\item{Boundary.knots}{- bounddary knots of the B-spline basis}

\item{ORDER}{- 1 step, 2 linear, 3 quadratic, 4 cubic}
}
\value{
B - matrix of basis of b-pline functions of the specified degree, evaluated at the "time" points
        B has dimensions length(time) x (number of interior points + degree + 1)
}
\description{
Wraper function to a B-spline basis calculation function in R (currently from package bspline2)
}
\details{
The current bSpline() function from the splines2 package extends the bs() function in the splines package for B-spline basis by allowing piecewise constant (left-closed and right-open except on the right boundary) spline basis of degree zero.
}
\examples{
# A basis of cubic B-splines with no interior points
B = generate_bspline_basis( time = 0:10, Interior.knots=c(), Boundary.knots=c(0,10) )
B
# Note that B has 4=3+1 functions (as many as the order of cubic B-splines)
# we can plot the basis using matplot
matplot( 0:10, B, type="l")

}
