% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline_regression_loglikelihood_fun.R
\name{srllikb_fun}
\alias{srllikb_fun}
\title{Objective function for b-spline hazard regression}
\usage{
srllikb_fun(par.alpha, yd, Wik, Zik)
}
\arguments{
\item{par.alpha}{- coefficients of a b-spline in a b-pline basis B (not provided)}

\item{yd}{- time-to-event data observations matrix, with each row being a pair (time,status)}

\item{Wik}{- matrix that multiplied by par.alpha gives the hazard function}

\item{Zik}{- matrix that multiplied by par.alpha gives the cumulative hazard function}
}
\value{
l - objective function (-2 *log likelihood function)
}
\description{
Calculates the objective function for b-spline hazard regression using as input the needed pre-calculated matrices
}
