% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etsim.R
\name{etsim}
\alias{etsim}
\title{Simulate time-to-event data}
\usage{
etsim(INPUTS)
}
\arguments{
\item{INPUTS}{- hazard and censoring distributions as generated by etsim_inputs()}
}
\value{
Returns a list of outputs (time, status, entry)

    t = array of times to event
    status = 0 if alive at time t, 1 otherwise
    entry = NULL [IS IT OK TO IGNORE LATE ENTRY AT THIS STAGE?]
}
\description{
Generates survival time and status fields using hazard and censoring as inputs
}
\details{
Times to event are defined as t = min( t_h, t_C), where t_h is the time to hazard (distributed according to the input S_h) and t_C is the time to censoring (distributed according to the input S_censoring)
}
\examples{
    #First, we generate input for a b-spline hazard simulation with piecewise exponential survival censoring
    knots = c(0, 0, 0, 0, 1, 3, 6, 10, 10, 10, 10)
    betac = 1 * c(0.05, 0.05, 0.05, 0.05, 0.40, 0.1, 0.05, NA, NA, NA, NA)
    HParm = data.frame(knots, betac) # 'A Simple B-Spline'
    cll = c(0, 5)
    cup = c(5, 10)
    cih = c(0.0125, 0.025)
    CParm = data.frame(cll, cup, cih) # 'Light Censoring'
    INPUTS = etsim_inputs( HParm=HParm, Cparm=Cparm)

    #Then, we can generate the time-to event data
    SimDat = etsim(INPUTS)
    table( SimDat$status )

}
