% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etsim_inputs.R
\name{etsim_inputs}
\alias{etsim_inputs}
\title{Generate hazard and censoring distributions}
\usage{
etsim_inputs(
  Hazard = "spline",
  HParam,
  Censor = "pe",
  CParam,
  Tmax = 10,
  SampleSize = 101
)
}
\arguments{
\item{Hazard}{- hazard function for outcome variable.
'exp' for exponential, 'weib' for Weibull, 'pe' for piecewise-exponential, or 'spline' for B-Spline}

\item{HParam}{- matrix of parameter values for outcome variable \cr}

\item{Censor}{- censoring hazard ('', 'exp', 'weib', 'pe' or 'bspline')}

\item{CParam}{- matrix of parameter values for censoring variable}

\item{Tmax}{- scalar, maximum follow-up time, default 10}

\item{SampleSize}{- scalar, sample size}
}
\value{
Returns a list of outputs (t, basis, h, hCensor, Sh, Scensor)

    t = array of times
    basis = basis of b-spline cubic functions [WE NEED TO ADD ORDER/DEGREE AS PARAMETER IF WE WANT TO ALLOW DIFFERENT B-SPLINES]
    h = simulated hazard distribution
    hCensor = simulated censoring distribution
    Sh = cumulative hazard survival function
    SCensor = cumulative censoring survival funtion
}
\description{
Generates hazard and censoring distributions and respective survival functions [IS IT OK TO IGNORE LATE ENTRY AT THIS STAGE?]
}
\examples{
#Generate input for a b-spline hazard simulation with piecewise exponential survival censoring
    knots = c(0, 1, 3, 6, 10, NaN, NaN )
    betac = 1 * c(0.05, 0.05, 0.05, 0.05, 0.40, 0.1, 0.05)
    HParm = data.frame(knots, betac) # 'A Simple B-Spline'
    cll = c(0, 5)
    cup = c(5, 10)
    cih = c(0.0125, 0.025)
    CParm = data.frame(cll, cup, cih) # 'Light Censoring'
    INPUTS = etsim_inputs( HParam=HParm, CParam=CParm )

}
