% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline_regression_objective_optimisation.R
\name{hazl_ker}
\alias{hazl_ker}
\title{Optimisation of the objective function}
\usage{
hazl_ker(yd, alpha0, Wik, Zik, Xh, XH)
}
\arguments{
\item{yd}{- data}

\item{alpha0}{- initial solution value for the B-spline coefficients}

\item{Wik}{- matrix as calculated by bspline_regression_basis_functions()}

\item{Zik}{- matrix as calculated by bspline_regression_basis_functions()}

\item{Xh}{- matrix as calculated by bspline_regression_basis_functions()}

\item{XH}{- matrix as calculated by bspline_regression_basis_functions()}
}
\value{
list(alpha1,hout,Sout,m2loglik) - alpha1 is the new solution;
using alpha1, the function also calculates the hazard (hout), survival (Sout) and objective function and gradient (m2loglik) ;
}
\description{
Optimisation of the objective function
}
