% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline_regression.R
\name{hspcore}
\alias{hspcore}
\title{Core function for B-spline hazard regression}
\usage{
hspcore(yd, ORDER = 4, knots, time, Bootstrap = 0, alphalevel = 0.95)
}
\arguments{
\item{yd}{- matrix of time and status data; y is event time, d is for delta, the status indicator}

\item{ORDER}{- 1 step, 2 linear, 3 quadratic, 4 cubic (i.e. ORDER=degree-1)}

\item{knots}{- optional sequence of knot locations including endpoints (without multiplicity)}

\item{time}{- vector of evaluation times}

\item{Bootstrap}{- number of bootstrap samples (set to zero if none)}

\item{alphalevel}{- alpha level for confidence intervals if bootstrap is being used}
}
\value{
list(alpha1,t,h,m2loglik,hb) - the optimal coefficients alpha1, the time variable, the optimal hazard, the objective function value and gradient, and the bootstrap hazards if any
}
\description{
This function allows to perform B-spline hazard regression given right-censored time-to-event data.
             If no b-spline knots and/or b-spline order are provided, automatic selection of those parameters is performed (see Details).
}
\details{
See \doi{<doi.org/10.2307/2532989>} the original paper by Philip S. Rosenberg.
}
